+++ .github/workflows/tests.yml
@@ -31,7 +31,7 @@ jobs:
       matrix:
         php-version: ['8.0', '8.1']
         connection: [mysql]
-        testsuite: [Api, Feature, Feature2, Commands, Unit-Models, Unit-Services]
+        testsuite: [Api, Feature, Commands-Other, Commands-Scheduling, Unit-Models, Unit-Services]
         # exclude:
         #   - php-version: '8.1'
         #     testsuite: Feature
