+++ phpunit.xml
@@ -28,12 +28,14 @@
       <directory suffix="Test.php">./tests/Feature</directory>
     </testsuite>
 
-    <testsuite name="Commands">
-      <directory suffix="Test.php">./tests/Commands</directory>
+    <testsuite name="Commands-Other">
+      <directory suffix="Test.php">./tests/Commands/OneTime</directory>
+      <directory suffix="Test.php">./tests/Commands/Other</directory>
     </testsuite>
 
-    <testsuite name="Feature2">
-      <directory suffix="Test.php">./tests/Feature2</directory>
+    <testsuite name="Commands-Scheduling">
+      <directory suffix="Test.php">./tests/Commands/Scheduling</directory>
+      <directory suffix="Test.php">./tests/Commands/Tests</directory>
     </testsuite>
 
     <testsuite name="Unit-Models">
