+++ tests/Commands/OneTime/MoveAvatarsToPhotosDirectoryTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Commands;
+namespace Tests\Commands\OneTime;
 
 use Tests\TestCase;
 use App\Models\User\User;
